+++ sdk/storage/Azure.Storage.Files.Shares/src/ShareFileClient.cs
@@ -4543,6 +4543,7 @@ namespace Azure.Storage.Files.Shares
         /// A <see cref="RequestFailedException"/> will be thrown if
         /// a failure occurs.
         /// </remarks>
+        [ForwardsClientCalls]
         public virtual Response<ShareFileUploadInfo> Upload(
             Stream stream,
             ShareFileUploadOptions options,
@@ -4583,6 +4584,7 @@ namespace Azure.Storage.Files.Shares
         /// A <see cref="RequestFailedException"/> will be thrown if
         /// a failure occurs.
         /// </remarks>
+        [ForwardsClientCalls]
         public virtual async Task<Response<ShareFileUploadInfo>> UploadAsync(
             Stream stream,
             ShareFileUploadOptions options,
@@ -4833,115 +4835,89 @@ namespace Azure.Storage.Files.Shares
             bool async,
             CancellationToken cancellationToken)
         {
-            Errors.VerifyStreamPosition(content, nameof(content));
-
-            // partitioned uploads don't support pre-calculated hashes
-            if (hashingOptions?.PrecalculatedHash != default)
-            {
-                throw Errors.PrecalculatedHashNotSupportedOnSplit();
-            }
-
-            DiagnosticScope scope = ClientConfiguration.ClientDiagnostics.CreateScope($"{nameof(ShareFileClient)}.{nameof(Upload)}");
-            try
-            {
-                scope.Start();
-
-                // Try to upload the file as a single range
-                Debug.Assert(singleRangeThreshold <= Constants.File.MaxFileUpdateRange);
-                var length = content?.Length - content?.Position;
-                if (length <= singleRangeThreshold)
+            var uploader = GetPartitionedUploader(
+                new StorageTransferOptions
                 {
-                    return await UploadRangeInternal(
-                        new HttpRange(0, length),
-                        content,
+                    // shares can't suppot parallel upload
+                    MaximumConcurrency = 1,
+                    MaximumTransferSize = singleRangeThreshold,
+                    InitialTransferSize = singleRangeThreshold
+                },
+                hashingOptions,
+                operationName: $"{nameof(ShareFileClient)}.{nameof(Upload)}");
+
+            return await uploader.UploadInternal(
+                content,
+                expectedContentLength: default,
+                new ShareFileUploadData
+                {
+                    Conditions = conditions,
+                },
+                progressHandler,
+                async,
+                cancellationToken)
+                .ConfigureAwait(false);
+        }
+        #endregion Upload
+
+        #region PartitionedUploader
+        internal class ShareFileUploadData
+        {
+            public ShareFileRequestConditions Conditions { get; set; }
+            public Response<ShareFileUploadInfo> LastUploadRangeResponse { get; set; }
+        }
+
+        internal PartitionedUploader<ShareFileUploadData, ShareFileUploadInfo> GetPartitionedUploader(
+            StorageTransferOptions transferOptions,
+            UploadTransactionalHashingOptions hashingOptions,
+            ArrayPool<byte> arrayPool = null,
+            string operationName = null)
+            => new PartitionedUploader<ShareFileUploadData, ShareFileUploadInfo>(
+                GetPartitionedUploaderBehaviors(this),
+                transferOptions,
+                hashingOptions,
+                arrayPool,
+                operationName);
+
+        // static because it makes mocking easier in tests
+        internal static PartitionedUploader<ShareFileUploadData, ShareFileUploadInfo>.Behaviors GetPartitionedUploaderBehaviors(ShareFileClient client)
+            => new PartitionedUploader<ShareFileUploadData, ShareFileUploadInfo>.Behaviors
+            {
+                SingleUpload = async (stream, data, progressHandler, hashingOptions, operationName, async, cancellationToken) =>
+                {
+                    return await client.UploadRangeInternal(
+                        new HttpRange(offset: 0, length: stream.Length),
+                        stream,
                         new ShareFileUploadRangeOptions
                         {
-                            TransactionalHashingOptions = hashingOptions,
+                            Conditions = data.Conditions,
                             ProgressHandler = progressHandler,
-                            Conditions = conditions
+                            TransactionalHashingOptions = hashingOptions,
                         },
                         async,
                         cancellationToken)
                         .ConfigureAwait(false);
-                }
-
-                // Otherwise naively split the file into ranges and upload them individually
-                var response = default(Response<ShareFileUploadInfo>);
-                var pool = default(MemoryPool<byte>);
-                // erase potential precalculated hash now that we're splitting; we'll have to recalculate.
-                hashingOptions = hashingOptions == default
-                    ? default
-                    : new UploadTransactionalHashingOptions
-                    {
-                        Algorithm = hashingOptions.Algorithm
-                    };
-
-                long initalPosition = content.Position;
-
-                try
+                },
+                UploadPartition = async (stream, offset, data, progressHandler, hashingOptions, async, cancellationToken) =>
                 {
-                    pool = (singleRangeThreshold < MemoryPool<byte>.Shared.MaxBufferSize) ?
-                        MemoryPool<byte>.Shared :
-                        new StorageMemoryPool(singleRangeThreshold, 1);
-                    for (; ; )
-                    {
-                        // Get the next chunk of content
-                        var parentPosition = content.Position;
-                        IMemoryOwner<byte> buffer = pool.Rent(singleRangeThreshold);
-                        if (!MemoryMarshal.TryGetArray<byte>(buffer.Memory, out ArraySegment<byte> segment))
+                    data.LastUploadRangeResponse = await client.UploadRangeInternal(
+                        new HttpRange(offset: offset, length: stream.Length),
+                        stream,
+                        new ShareFileUploadRangeOptions
                         {
-                            throw Errors.UnableAccessArray();
-                        }
-                        var count = async ?
-                            await content.ReadAsync(segment.Array, 0, singleRangeThreshold, cancellationToken).ConfigureAwait(false) :
-                            content.Read(segment.Array, 0, singleRangeThreshold);
-
-                        // Stop when we've exhausted the content
-                        if (count <= 0)
-                        { break; }
-
-                        // Upload the chunk
-                        var partition = new StreamPartition(
-                            buffer.Memory,
-                            parentPosition,
-                            count,
-                            () => buffer.Dispose(),
-                            cancellationToken);
-                        response = await UploadRangeInternal(
-                            new HttpRange(partition.ParentPosition - initalPosition, partition.Length),
-                            partition,
-                            new ShareFileUploadRangeOptions
-                            {
-                                TransactionalHashingOptions = hashingOptions,
-                                ProgressHandler = progressHandler,
-                                Conditions = conditions
-                            },
-                            async,
-                            cancellationToken)
-                            .ConfigureAwait(false);
-                    }
-                }
-                finally
-                {
-                    if (pool is StorageMemoryPool)
-                    {
-                        pool.Dispose();
-                    }
-                }
-                return response;
-            }
-            catch (Exception ex)
-            {
-                ClientConfiguration.Pipeline.LogException(ex);
-                scope.Failed(ex);
-                throw;
-            }
-            finally
-            {
-                scope.Dispose();
-            }
-        }
-        #endregion Upload
+                            Conditions = data.Conditions,
+                            ProgressHandler = progressHandler,
+                            TransactionalHashingOptions = hashingOptions,
+                        },
+                        async,
+                        cancellationToken)
+                        .ConfigureAwait(false);
+                },
+                CommitPartitionedUpload = (partitions, data, async, cancellationToken) => Task.FromResult(data.LastUploadRangeResponse),
+                Scope = operationName => client.ClientConfiguration.ClientDiagnostics.CreateScope(operationName ??
+                    $"{nameof(Azure)}.{nameof(Storage)}.{nameof(Files)}.{nameof(Shares)}.{nameof(ShareFileClient)}.{nameof(ShareFileClient.Upload)}")
+            };
+        #endregion
 
         #region GetRangeList
         /// <summary>