+++ sdk/storage/Azure.Storage.Files.Shares/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 12.9.0-beta.3 (Unreleased)
 - Added support for service version 2021-04-10.
+- Fixed a memory leak in ShareFileClient.UploadAsync().
 
 ## 12.9.0-beta.2 (2021-11-30)
 - Added support for service version 2021-02-12.